feat: root group and react suspense
