+++ app/dashboard/(overview)/loading.tsx
@@ -1,4 +1,5 @@
-import DashboardSkeleton from "../ui/skeletons";
+import DashboardSkeleton from "@/app/ui/skeletons";
+
 
 export default function Loading() {
   return <DashboardSkeleton />
